+++ src/components/common/buttons/primary-button/index.tsx
@@ -1,27 +1,9 @@
-import { ButtonTag, StyleProps } from '@/types/components/common/button.ts'
-import { handleButtonStyleProps } from '@/utils/button-props.ts'
+import { PrimaryButtonTag } from '@/types/components/common/button.ts'
 
-export const PrimaryButton = ({
-    dataProps,
-    styleProps,
-    index,
-}: {
-    dataProps: ButtonTag
-    styleProps?: StyleProps
-    index: number
-}) => {
-    const activeFilter = {
-        ...styleProps,
-        font: dataProps.isActive ? 'type2' : 'type1',
-    } as StyleProps
+export const PrimaryButton = ({ onClick, type, style, label }: PrimaryButtonTag) => {
     return (
-        <button
-            onClick={() => dataProps?.onClick?.(index)}
-            type={dataProps?.type || 'button'}
-            aria-label={dataProps?.isActive ? 'isActive' : 'none'}
-            className={handleButtonStyleProps(activeFilter)}
-        >
-            {dataProps.label}
+        <button onClick={onClick} type={type || 'button'} className={style}>
+            {label}
         </button>
     )
 }
